requirements.txt
boto3==1.37.28
openai==1.66.3
pandas==2.2.3
praw==7.8.1
python-dotenv==1.1.0
requests==2.32.3
ruff==0.11.0